Fix extraction widgets reset upon save csv

+++ sql/core/inventory/000_create_tables.sql
@@ -25,7 +25,7 @@ CREATE TABLE session_batch_inventory (
     session_id                  INT     NOT NULL,
     batch_inventory_id          INT     NOT NULL,
 	production_date             DATE,
-    quantity_used               INT     NOT NULL CHECK (quantity_used >= 1),
+    quantity_used               NUMERIC(7,2)     NOT NULL CHECK (quantity_used >= 1),
     batch_code                  TEXT,
     unit                        TEXT    NOT NULL CHECK (unit IN ('pcs', 'g')),
 	UNIQUE (session_id, batch_inventory_id)
